관심있는 식당 정보 저장하기

+++ src/main/java/com/github/yumyum/map/web/dto/restaurant/RestaurantDTO.java
@@ -15,7 +15,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class RestaurantDTO {
-    private Long rstrId;
+    private Integer rstrId;
     private String rstrName;
     private String address;
     private String latitude;
